Rename and move functions to reflect standards
'onUpdate' and 'Draw' seem to be default names for the
functionality we're using here.

I left some code in sceneOneBeach.go as I feel there could
go scene specific functions later.
